requirements.txt
# Data science & ML
pandas
numpy
scikit-learn
matplotlib
seaborn
scipy

# Deep learning
tensorflow
torch
transformers

# API
fastapi
uvicorn
requests
python-dotenv
flask
flask-socketio
eventlet
psycopg2-binary
